Remove code to show S3 link in search results.
The code to emit the link now first calls S3 to see if the item has any files in Item Storage and if not, hides the link. It's an expensive operation to be performing on every item in the search results. Instead, you can see the S3 link in the admin view for the item.